updated pie charts

+++ src/Models/SummaryModels.cs
@@ -17,9 +17,9 @@ namespace CashTrack.Models.SummaryModels
         public List<SubCategoryQuickView> TopCategories { get; set; }
         public List<MerchantQuickView> TopMerchants { get; set; }
         public List<IncomeSourceQuickView> TopSources { get; set; }
-        public Dictionary<string, decimal> SubCategoryPercentages { get; set; }
+        public Dictionary<string, int> SubCategoryPercentages { get; set; }
         public Dictionary<string, decimal> MainCategoryPercentages { get; set; }
-        public Dictionary<string, decimal> MerchantPercentages { get; set; }
+        public Dictionary<string, int> MerchantPercentages { get; set; }
         public Dictionary<string, decimal> IncomeSourcePercentages { get; set; }
         public List<TransactionBreakdown> TransactionBreakdown { get; set; }
         public List<AnnualStatistics> ExpenseStatistics { get; set; }
@@ -88,9 +88,9 @@ namespace CashTrack.Models.SummaryModels
         public AnnualSavingsChart SavingsChart { get; set; }
         public AnnualIncomeExpenseChart IncomeExpenseChart { get; set; }
         public AnnualSummaryTotals AnnualSummary { get; set; }
-        public Dictionary<string, decimal> SubCategoryPercentages { get; set; }
+        public Dictionary<string, int> SubCategoryPercentages { get; set; }
         public Dictionary<string, decimal> MainCategoryPercentages { get; set; }
-        public Dictionary<string, decimal> MerchantPercentages { get; set; }
+        public Dictionary<string, int> MerchantPercentages { get; set; }
         public Dictionary<string, decimal> IncomeSourcePercentages { get; set; }
         public List<MonthlyStatistics> MonthlyExpenseStatistics { get; set; }
         public AnnualMonthlySummaryChart AnnualMonthlySummaryChart { get; set; }
@@ -145,9 +145,9 @@ namespace CashTrack.Models.SummaryModels
         public ExpenseSummaryChartData ExpenseSummaryChart { get; set; }
         public MonthlySummary MonthlySummary { get; set; }
         public OverallSummaryChart OverallSummaryChart { get; set; }
-        public Dictionary<string, decimal> SubCategoryPercentages { get; set; }
+        public Dictionary<string, int> SubCategoryPercentages { get; set; }
         public Dictionary<string, decimal> MainCategoryPercentages { get; set; }
-        public Dictionary<string, decimal> MerchantPercentages { get; set; }
+        public Dictionary<string, int> MerchantPercentages { get; set; }
         public DailyExpenseChart DailyExpenseLineChart { get; set; }
         public MonthlyYearToDate YearToDate { get; set; }
         public List<ExpenseQuickView> TopExpenses { get; set; }
